requirements.txt
flask==2.3.3
openai==1.3.3
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.5.0
pydantic==2.4.2
twilio==8.5.0
python-dateutil==2.8.2 